Pjoter

+++ app/src/main/java/com/example/piggerbank/Baza/Money.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 data class Money(
     @PrimaryKey(autoGenerate = false)
     @ColumnInfo val id: Int?,
-    @ColumnInfo val moneyValue: Float,
+    @ColumnInfo val moneyValue: Double,
     @ColumnInfo val moneyDescription: String?,
     @ColumnInfo val moneyCategory_id: Int?
 )
